+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **added**: Components (including examples for them)
   - `component:IconButton`
   - `component:Button`
+  - `component:Checkbox`
   - `component:Icon`
   - `component:RadioButton`
   - `component:StateLayer`